Create Block changes between web1, web2, and web3
BREAKING CHANGES BETWEEN WEB 1.0, 2.0 AND 3.0

Firstly, Web 1.0 was specifically for Reading as well as getting information. Web 2.0 included Reading, Writing and Creating information. Furthermore, We3.0 incorporates Reading, Writing as well as Owning.

Secondly, Web1.0 is used in Dictionaries, Web forms and it uses HTML portals, while Web2.0 is used for website applications and it uses XML and RSS platforms. And Web3.0 is used for smart contracts applications and it uses RDF, RDFS, and OWL.

In addition, Web1.0 engages in Banner Advertising, while Web2.0 engages in Interactive advertising. And Web3.0 deals with Behavioural advertising.

Lastly, Web1.0 uses Content Delivery Network which showcases the piece of information on the website. Web 2.0 uses Ajax and JavaScript frameworks for it’s development. While in Web3.0, every Device is connected to the Website using ontological languages like OWL